reworked on module of js
