+++ index.html
@@ -64,7 +64,7 @@
       </nav>
 
       <div class="cart-design-container">
-        <h2 class="cart-heading">Your Cart</h2>
+        <h2 class="cart-heading"></h2>
         <ul class="cart-info"></ul>
         <div class="checkout-price-info"></div>
         <div class="cart-buttons"></div>
